Split Ribbon view from Ribbon controller
The View of the Ribbon is now contained in its own file inside the views
package.
