+++ src/main/java/gui/controllers/ribbon/RibbonController.java
@@ -1,12 +1,6 @@
 package gui.controllers.ribbon;
 
-import java.io.File;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.LinkedList;
-import java.util.List;
-import java.util.Queue;
 import java.util.ResourceBundle;
 
 import javafx.beans.value.ChangeListener;
@@ -20,13 +14,10 @@ import javafx.scene.control.ScrollPane;
 import javafx.scene.input.KeyEvent;
 import javafx.scene.input.ScrollEvent;
 import javafx.scene.layout.GridPane;
-import javafx.scene.paint.Paint;
-import javafx.scene.shape.Line;
 
-import parsers.XlsxParser;
 import db.DatabaseManager;
 import gui.Launcher;
-import gui.views.phylogeny.NewickColourMatching;
+import gui.views.ribbon.RibbonView;
 
 /**
  * Controller class for the Ribbon screen/tab.
@@ -42,19 +33,17 @@ public class RibbonController implements Initializable {
 	@FXML private CheckBox checkboxSnp;
 	@FXML private CheckBox checkboxInsert;
 	
-	private DatabaseManager dbm = Launcher.getDatabaseManager();
-	
 	private Group innerGroup;
 	private Group outerGroup;
 	private Group otherGroup;
 	private Group annotationRibbonGroup;
 	private Group annotationGraphGroup;
-
-
-	private HashMap<String, String> lineages = updateLineages();
-	private ArrayList<Integer> genomeIds = createList();
-	private Group collapsedGroup = createCollapsedRibbons();
-	private Group normalGroup = createNormalRibbons();
+	
+	private DatabaseManager dbm = Launcher.getDatabaseManager();
+	private RibbonView ribbonView = new RibbonView(dbm);
+	
+	private Group collapsedGroup = ribbonView.createCollapsedRibbons();
+	private Group normalGroup = ribbonView.createNormalRibbons();
 	
     private static final double MAX_SCALE = 1.0d;
     private static final double MIN_SCALE = .003d;
@@ -63,12 +52,7 @@ public class RibbonController implements Initializable {
 	
 	private double prevScale = 1;
 	
-	/**
-	 * HashMap containing the lineages of the specimens.
-	 */
-	
-	private static String xlsxpath = System.getProperty("user.dir") + File.separator + "Data"
-			+ File.separator + "TB10" + File.separator + "metadata" + ".xlsx";
+
 	
 	/**
 	 * Handles the scroll wheel event for the ribbon view.
@@ -226,11 +210,7 @@ public class RibbonController implements Initializable {
 		
 	}
 	
-	private HashMap<String, String> updateLineages() {
-		XlsxParser xlsxparser = new XlsxParser();
-		xlsxparser.parse(xlsxpath);
-		return xlsxparser.getLineages();
-	}
+
 	
 	/**
 	 * Updates the view. Used when changing database files so the graph
@@ -242,155 +222,10 @@ public class RibbonController implements Initializable {
 		outerGroup = new Group(innerGroup);
 		scrollPane.setContent(outerGroup);
 	}
-	
-
-	/**
-	 * Creates all paths that make up the ribbons and returns a {@link Group}
-	 * containing those paths.
-	 * 
-	 * @return A group containing the ribbons.
-	 */
-	public Group createNormalRibbons() {
-		System.out.println("Creating normal ribbons");
-		Group res = new Group();
-		ArrayList<ArrayList<Integer>> links = dbm.getDbReader().getLinks(genomeIds);
-		ArrayList<ArrayList<Integer>> counts = dbm.getDbReader().getLinkWeights(genomeIds);
-		ArrayList<ArrayList<Paint>> colours = calculateColours(links, genomeIds);
-		ArrayList<Integer> xcoords = dbm.getDbReader().getAllXCoord();
-		ArrayList<Integer> ycoords = dbm.getDbReader().getAllYCoord();
-		
-		for (int fromId = 1; fromId <= links.size(); fromId++) {
-			for (int j = 0; j < links.get(fromId - 1).size(); j++) {
-				int toId = links.get(fromId - 1).get(j);
-				Line line = new Line(xcoords.get(fromId - 1), ycoords.get(fromId - 1), 
-						xcoords.get(toId - 1), ycoords.get(toId - 1));
-				line.setStrokeWidth(counts.get(fromId - 1).get(j));
-				line.setStroke(colours.get(fromId - 1).get(j));
-		        res.getChildren().add(line);
-			}
-		}
-		System.out.println("Finished normal ribbons");
-		return res;
-	}
-	
-	public Paint getLineColor(int ff, int tt) {
-		Paint color = Paint.valueOf("0xff0000ff");
-		ArrayList<String> from = dbm.getDbReader().getGenomesThroughSegment(ff);
-		ArrayList<String> to = dbm.getDbReader().getGenomesThroughSegment(tt);
-		if (from.size() > to.size()) {
-			for (int i = 0; i < to.size(); i++) {
-				String genome = to.get(i);
-				if (lineages.containsKey(genome) && from.contains(genome) 
-						&& !genome.equals("MT_H37RV_BRD_V5.ref")) {
-					return NewickColourMatching.getLineageColour(lineages.get(genome));
-				}
-			}
-		} else if (from.size() < to.size()) {
-			for (int i = 0; i < from.size(); i++) {
-				String genome = from.get(i);
-				if (lineages.containsKey(genome) && to.contains(genome) 
-						&& !genome.equals("MT_H37RV_BRD_V5.ref")) {
-					return NewickColourMatching.getLineageColour(lineages.get(genome));
-				}
-			}
-		} else {
-			for (int i = 0; i < from.size(); i++) {
-				String genome = from.get(i);
-				if (lineages.containsKey(genome) && to.contains(genome) 
-						&& !genome.equals("MT_H37RV_BRD_V5.ref")) {
-					return NewickColourMatching.getLineageColour(lineages.get(genome));
-				}
-			}
-		}
-		return color;
-	}
-	
-	private ArrayList<ArrayList<Paint>> calculateColours(ArrayList<ArrayList<Integer>> linkIds, 
-			ArrayList<Integer> genomes) {
-		ArrayList<ArrayList<Paint>> colours = 
-				new ArrayList<ArrayList<Paint>>();
-		for (int i = 0; i < dbm.getDbReader().countSegments(); i++) {
-			colours.add(new ArrayList<Paint>());
-		}
-		ArrayList<String> genomeNames = dbm.getDbReader().getGenomeNames(genomes);
-		
-		HashMap<Integer, ArrayList<Integer>> hash = dbm.getDbReader().getGenomesPerLink(genomes);
-		for (int i = 0; i < linkIds.size(); i++) {
-			for (int j = 0; j < linkIds.get(i).size(); j++) {
-				ArrayList<Integer> genomeIds = hash.get(100000 * (i + 1) 
-						+ linkIds.get(i).get(j));
-				int id = genomeIds.get(0);
-				Paint colour = Paint.valueOf("0xff0000ff");
-				String genome = genomeNames.get(id - 1);
-				if (!genome.startsWith("M")) {
-					colour = NewickColourMatching
-							.getLineageColour(lineages.get(genome));
-				} 
-				colours.get(i).add(colour);
-			}
-		}
-		return colours;
-	}
-	
-	/**
-	 * Creates all paths that make up the collapsed ribbons and returns a
-	 * {@link Group} containing those paths.
-	 * 
-	 * @return A group containing the ribbons.
-	 */
-	public Group createCollapsedRibbons() {
-		System.out.println("Creating collapsed ribbons");
-		Group res = new Group();
-		ArrayList<ArrayList<Integer>> links = dbm.getDbReader().getLinks(genomeIds);
-		ArrayList<ArrayList<Integer>> counts = dbm.getDbReader().getLinkWeights(genomeIds);
-		ArrayList<ArrayList<Paint>> colours = calculateColours(links, genomeIds);
-		ArrayList<Integer> xcoords = dbm.getDbReader().getAllXCoord();
-		ArrayList<Integer> ycoords = dbm.getDbReader().getAllYCoord();
-		Queue<int[]> bubbles = new LinkedList<>(dbm.getDbReader().getBubbles(genomeIds));
-		
-		List<Integer> ignore = new LinkedList<>();
-		
-		for (int fromId = 1; fromId <= links.size(); fromId++) {
-			List<Integer> edges = links.get(fromId - 1);
-			
-			if (!bubbles.isEmpty() && fromId == bubbles.peek()[0]) {
-				int[] bubble = bubbles.poll();
-				Line line = new Line(xcoords.get(fromId - 1), ycoords.get(fromId - 1), 
-						xcoords.get(bubble[1] - 1), ycoords.get(bubble[1] - 1));
-				double width = bubble[2];
-				line.setStrokeWidth(2 * width);
-				line.setStroke(colours.get(fromId - 1).get(0));
-		        res.getChildren().add(line);
-		        ignore.addAll(edges);
-			} else {
-				if (ignore.contains(fromId)) {
-					continue;
-				}
-				for (int toId : edges) {
-					if (!bubbles.isEmpty() && toId == bubbles.peek()[1]) {
-						break;
-					}
-
-					for (int j = 0; j < links.get(fromId - 1).size(); j++) {
-						Line line = new Line(xcoords.get(fromId - 1), 
-								ycoords.get(fromId - 1), 
-								xcoords.get(toId - 1), 
-								ycoords.get(toId - 1));
-						line.setStrokeWidth(2 * counts.get(fromId - 1).get(j));
-						line.setStroke(colours.get(fromId - 1).get(j));
-						res.getChildren().add(line);
-					}
-				}
-			}
-		}
-		System.out.println("Finished collapsed ribbons");
-		
-		return res;
-	}
 
 	public void redraw() {
-		collapsedGroup = createCollapsedRibbons();
-		normalGroup = createNormalRibbons();
+		collapsedGroup = ribbonView.createCollapsedRibbons();
+		normalGroup = ribbonView.createNormalRibbons();
 		updateView();
 		double maxY = dbm.getDbReader().getMaxYCoord();
 		innerGroup.setScaleY(720.0 / maxY);
@@ -428,16 +263,12 @@ public class RibbonController implements Initializable {
 	public void setAnnotationGraphPane(ScrollPane scrollpane) {
 		annotationGraphPane = scrollpane;
 	}
-	
-	public void setGenomeIds(ArrayList<Integer> ids) {
-		genomeIds = ids;
+
+
+
+	public RibbonView getRibbonView() {
+		return ribbonView;
 	}
 	
-	public ArrayList<Integer> createList() {
-		ArrayList<Integer> list = new ArrayList<Integer>();
-		for (int i = 0; i < 10; i++) {
-			list.add(i);
-		}
-		return list;
-	}
+
 }
